// frontend/src/lib/api/generated/schemas/appealListItemResponsePriority.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Baltiyskiy Bereg Backend API
 * Backend API for the Baltiyskiy Bereg service-desk assistant. It exposes admin endpoints for dashboards and support operators, plus public client endpoints shared by the web flow and Telegram bot.
 * OpenAPI spec version: 1.0.0
 */

export type AppealListItemResponsePriority = typeof AppealListItemResponsePriority[keyof typeof AppealListItemResponsePriority];


export const AppealListItemResponsePriority = {
  p1: 'p1',
  p2: 'p2',
  p3: 'p3',
  p4: 'p4',
} as const;